adding new branch develop for redux part
